fix: `EvaluationAgent` metric computation bugs (#6)
* fix: `GPUMemory` as string

* fix: return format of evaluation results (no tensors)

* fix: inf evaluation results for transformers


* feat: raise error when device movement fails

* fix: add exceptions to handler registration


---------

+++ src/pruna/engine/handler/handler_transformer.py
@@ -67,7 +67,7 @@ class TransformerHandler(InferenceHandler):
         Any
             The processed output.
         """
-        return output.logits
+        return output.logits.float()
 
     def log_model_info(self) -> None:
         """Log information about the inference handler."""
